Feature DPoS - Adjust file structure
