+++ AElf.Launcher/ChainInfo.json
@@ -1,1 +1,1 @@
-{"id":"0xe68cc01b9f796496cde27c7b30a921b2e9a8e0a2f6cc5864db4c127abfefb8c9"}
+{"id":"0x5c52962ab99d487a42b30ffad222d543734a325ace74858d4318a162e8b97e7a"}
